src/requirements.txt
pytesseract
easyocr
pypdf
pdfminer.six
python-docx
google-generativeai